Added CCD. Fixed a bug in the frame width calculation. Modified cone vector generator to work in degrees, not radians (raysect angles are all in degrees)
